Fitbit OAuth 2.0 Authorization
I found a good example of a fitbit implementation w/ Oauth 2.0 here: https://github.com/Stasonis/fitbit-api-example-android

I only used the code that was necessary to get the authorization to work, but I think that there is also a good example of how to use the data.

Currently, the app is hardcoded to use the client id, secret, and redirect url from our fitbit account, but I believe that it does login properly.
After logging in, the app does nothing, but it shouldn't be too hard to figure that stuff out.